Make PeopleEventsProvider an interface and rename the implementation CompositeProvider
