+++ android_mobile/src/main/java/com/alexstyl/specialdates/people/PeopleModule.java
@@ -2,7 +2,7 @@ package com.alexstyl.specialdates.people;
 
 import com.alexstyl.specialdates.CrashAndErrorTracker;
 import com.alexstyl.specialdates.Strings;
-import com.alexstyl.specialdates.events.peopleevents.PeopleEventsProvider;
+import com.alexstyl.specialdates.events.peopleevents.CompositePeopleEventsProvider;
 import com.alexstyl.specialdates.facebook.FacebookUserSettings;
 import com.alexstyl.specialdates.permissions.MementoPermissions;
 
@@ -21,7 +21,7 @@ public class PeopleModule {
 
     @Provides
     PeoplePresenter presenter(MementoPermissions permissions,
-                              PeopleEventsProvider peopleEventsProvider,
+                              CompositePeopleEventsProvider peopleEventsProvider,
                               PeopleViewModelFactory viewModelFactory,
                               CrashAndErrorTracker errorTracker) {
         return new PeoplePresenter(
