+++ android_mobile/src/main/java/com/alexstyl/specialdates/upcoming/widget/today/QueryUpcomingPeopleEventsTask.java
@@ -3,15 +3,15 @@ package com.alexstyl.specialdates.upcoming.widget.today;
 import android.os.AsyncTask;
 
 import com.alexstyl.specialdates.Optional;
+import com.alexstyl.specialdates.events.peopleevents.CompositePeopleEventsProvider;
 import com.alexstyl.specialdates.events.peopleevents.ContactEventsOnADate;
 import com.alexstyl.specialdates.date.Date;
-import com.alexstyl.specialdates.events.peopleevents.PeopleEventsProvider;
 
 abstract class QueryUpcomingPeopleEventsTask extends AsyncTask<Object, Object, Optional<ContactEventsOnADate>> {
 
-    private final PeopleEventsProvider eventsProvider;
+    private final CompositePeopleEventsProvider eventsProvider;
 
-    QueryUpcomingPeopleEventsTask(PeopleEventsProvider eventsProvider) {
+    QueryUpcomingPeopleEventsTask(CompositePeopleEventsProvider eventsProvider) {
         this.eventsProvider = eventsProvider;
     }
