+++ memento/src/main/java/com/alexstyl/specialdates/upcoming/CompositeUpcomingEventsProvider.kt
@@ -7,9 +7,9 @@ import com.alexstyl.specialdates.events.bankholidays.BankHolidaysPreferences
 import com.alexstyl.specialdates.events.namedays.NamedayUserSettings
 import com.alexstyl.specialdates.events.namedays.NamesInADate
 import com.alexstyl.specialdates.events.namedays.calendar.resource.NamedayCalendarProvider
-import com.alexstyl.specialdates.events.peopleevents.PeopleEventsProvider
+import com.alexstyl.specialdates.events.peopleevents.CompositePeopleEventsProvider
 
-class CompositeUpcomingEventsProvider(private val peopleEventsProvider: PeopleEventsProvider,
+class CompositeUpcomingEventsProvider(private val peopleEventsProvider: CompositePeopleEventsProvider,
                                       private val namedayPreferences: NamedayUserSettings,
                                       private val namedayCalendarProvider: NamedayCalendarProvider,
                                       private val bankHolidaysPreferences: BankHolidaysPreferences,
